Move DateInput widget
