+++ src/components/DateInput/DateInput.js
@@ -19,9 +19,9 @@
 import React from 'react';
 import { observer } from 'mobx-react';
 import { observable, computed, action } from 'mobx';
-import { ActionButtonFill } from './ActionButtons';
+import { ActionButtonFill } from '../ActionButtons/ActionButtons';
 import ModalPortalWrapper from '../ModalPortalWrapper';
-import DatePicker from '../DatePicker/DatePicker';
+import DatePicker from './DatePicker';
 
 const re = new RegExp('(^-?[0-9]*)/?([0-9]*)/?([0-9]*)');
