+++ src/containers/AdminSTV/AdminSTV.js
@@ -20,7 +20,6 @@ import React from 'react';
 import { inject, observer } from 'mobx-react';
 import { computed, observable, action } from 'mobx';
 import { withRouter } from 'react-router-dom';
-import SmoothCollapse from 'react-smooth-collapse';
 
 import ActionButton, { CreateActionButton } from '../../components/ActionButtons/ActionButtons';
 import AdminWrapper from '../../components/AdminWrapper/AdminWrapper';
@@ -29,8 +28,7 @@ import AdminSTVCard from './AdminSTVCard';
 import './AdminSTV.less';
 import TwoActions from '../../components/TwoActions/TwoActions';
 import AdminTECard from '../AdminTE/AdminTECard';
-import UploadWidget from '../../components/ActionButtons/UploadWidget';
-import DateInput from '../../components/ActionButtons/DateInput';
+import AdminSTVAdd from './AdminSTVAdd';
 
 const STVTableHeader = () => (
   <div className='tooltip-author stv-entity--grid'>
@@ -82,19 +80,7 @@ class AdminSTV extends React.Component {
             : <ActionButton text='Back' icon='exit' click={() => this.props.history.push('/admin/te/')} />}
           <ActionButton text='Save' icon='save' click={() => null} />
         </TwoActions>
-        <SmoothCollapse expanded={this.add}>
-          <div className='stv-entity stv-entity--grid stv-entity--new'>
-            <div style={{ gridArea: '1 / 1 / 1 / 3' }}>
-              Start Date
-              <DateInput />
-              End Date
-              <DateInput />
-            </div>
-            <div style={{ gridArea: '1 / 3 / 1 / 5' }}>
-              <UploadWidget />
-            </div>
-          </div>
-        </SmoothCollapse>
+        <AdminSTVAdd add={this.add} />
         <TwoActions>
           <></>
           {this.add
